% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigFimo.R
\name{subdivideGenomicRegion}
\alias{subdivideGenomicRegion}
\title{given chrom, start, end, and a number of regions, subdivide with overlaps}
\usage{
subdivideGenomicRegion(chrom, start, end, count, overlap = 30)
}
\arguments{
\item{chrom}{character chromosome}

\item{start}{numeric of region}

\item{end}{numeric of region}

\item{count}{numeric the number to calculate}

\item{overlap}{numeric of neighboring regions, also used to extend start and end}
}
\value{
a data.frame (chrom, start, end, size) of approximately equal count regions
}
\description{
given chrom, start, end, and a number of regions, subdivide with overlaps
}
