% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigFimo.R
\name{BigFimo}
\alias{BigFimo}
\title{BigFimo}
\description{
An R6 Class to prepare for, and then run, multiple instances of FIMO.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BigFimo$new()}}
\item \href{#method-queryGeneHancer}{\code{BigFimo$queryGeneHancer()}}
\item \href{#method-get.tbl.gh}{\code{BigFimo$get.tbl.gh()}}
\item \href{#method-get.tbl.gh.oc}{\code{BigFimo$get.tbl.gh.oc()}}
\item \href{#method-createMemeFile}{\code{BigFimo$createMemeFile()}}
\item \href{#method-calculateRegionsForFimo}{\code{BigFimo$calculateRegionsForFimo()}}
\item \href{#method-createIndicesToDistributeTasks}{\code{BigFimo$createIndicesToDistributeTasks()}}
\item \href{#method-createFimoTables}{\code{BigFimo$createFimoTables()}}
\item \href{#method-getFimoRegionsFileList}{\code{BigFimo$getFimoRegionsFileList()}}
\item \href{#method-runMany}{\code{BigFimo$runMany()}}
\item \href{#method-clone}{\code{BigFimo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new BigFimo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$new(
  targetGene,
  genome = "hg38",
  tbl.oc,
  processCount,
  fimoThreshold,
  use.genehancer = TRUE,
  gh.elite.only = TRUE,
  maxGap.between.oc.and.gh = 5000,
  ocExpansion = 100,
  chrom = NA,
  start = NA,
  end = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetGene}}{character, the gene of interest.}

\item{\code{genome}}{character default "hg38"}

\item{\code{tbl.oc}}{data.frame, specified areas of (additional) open chromatin in which to run fio}

\item{\code{processCount}}{integer number of parallel processes}

\item{\code{fimoThreshold}}{numeric, 1e-4 for example}

\item{\code{use.genehancer}}{logical, use GeneHancer for open chromatin, default TRUE}

\item{\code{gh.elite.only}}{logical, use only doubly-attested GeneHancer regions, default TRUE}

\item{\code{maxGap.between.oc.and.gh}}{numeric, 5000 by default}

\item{\code{ocExpansion}}{numeric, pad reported open chromatin by this amount, default 100}

\item{\code{chrom}}{character optional fimo region, used rather than fimo & tbl.oc? default NA}

\item{\code{start}}{numeric optional fimo region, used rather than fimo & tbl.oc? default NA}

\item{\code{end}}{numeric optional fimo region, used rather than fimo & tbl.oc? default NA}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `BigFimo` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queryGeneHancer"></a>}}
\if{latex}{\out{\hypertarget{method-queryGeneHancer}{}}}
\subsection{Method \code{queryGeneHancer()}}{
call the genehancer database on khaleesi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$queryGeneHancer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.tbl.gh"></a>}}
\if{latex}{\out{\hypertarget{method-get.tbl.gh}{}}}
\subsection{Method \code{get.tbl.gh()}}{
access to the previously extracted genehancer table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$get.tbl.gh()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.tbl.gh.oc"></a>}}
\if{latex}{\out{\hypertarget{method-get.tbl.gh.oc}{}}}
\subsection{Method \code{get.tbl.gh.oc()}}{
access to the previously calculated overlap of genehancer and open chromatin tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$get.tbl.gh.oc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createMemeFile"></a>}}
\if{latex}{\out{\hypertarget{method-createMemeFile}{}}}
\subsection{Method \code{createMemeFile()}}{
fimo needs a meme file with motifs of interest.  just jaspar2022 for now
used to be (before june 2022)    jaspar2018 & hocomoco-core-A
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$createMemeFile()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateRegionsForFimo"></a>}}
\if{latex}{\out{\hypertarget{method-calculateRegionsForFimo}{}}}
\subsection{Method \code{calculateRegionsForFimo()}}{
find the overlap of oc with genehancer - but very generously
any open chromatin within < maxGap.between.oc.and.gh (often 5000)
 will be included.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$calculateRegionsForFimo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame the overlap of open chromatin with genehancer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createIndicesToDistributeTasks"></a>}}
\if{latex}{\out{\hypertarget{method-createIndicesToDistributeTasks}{}}}
\subsection{Method \code{createIndicesToDistributeTasks()}}{
we want to partition the fimo regions equally across the approximate
number of suggested processes.  this function returns a list of indices
each a vector. sometimes an extra process is needed, sometimes fewer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$createIndicesToDistributeTasks(tbl, suggestedProcessCount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{data.frame, specifies regions we want to distribute across multiple processes}

\item{\code{suggestedProcessCount}}{integer, may be adjusted up or down}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of integer vectors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createFimoTables"></a>}}
\if{latex}{\out{\hypertarget{method-createFimoTables}{}}}
\subsection{Method \code{createFimoTables()}}{
create one binary data.frame per process, splitting regions equally among them
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$createFimoTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector, the list of files containing the data.frames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFimoRegionsFileList"></a>}}
\if{latex}{\out{\hypertarget{method-getFimoRegionsFileList}{}}}
\subsection{Method \code{getFimoRegionsFileList()}}{
the files listed here specify regions in which fimo will search
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$getFimoRegionsFileList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runMany"></a>}}
\if{latex}{\out{\hypertarget{method-runMany}{}}}
\subsection{Method \code{runMany()}}{
initiates one fimo processes per region file, each previously created
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$runMany()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigFimo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
